recruit/requirements.txt
amqp==5.2.0
appnope==0.1.3
asgiref==3.8.1
asttokens==2.4.1
billiard==4.2.0
blinker==1.7.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.0
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
Django==5.0.7
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
executing==2.0.1
Flask==3.0.0
fonttools==4.51.0
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
kombu==5.3.7
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.4
openpyxl==3.1.5
packaging==23.2
pandas==2.2.2
parso==0.8.3
pexpect==4.8.0
pillow==10.3.0
platformdirs==4.0.0
prompt_toolkit==3.0.47
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.1
redis==5.0.7
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
threadpoolctl==3.2.0
tornado==6.3.3
traitlets==5.13.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
